database/requirements.txt
Django>=4.2,<5.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0

# AI services — these versions install cleanly together on Python 3.11–3.14.
torch>=2.2
transformers>=4.44,<5.0
tokenizers>=0.19,<1.0
sentence-transformers>=2.7,<3.0
huggingface_hub>=0.24,<1.0
keybert>=0.8,<1.0
pypdf>=4.0,<6.0
sentencepiece>=0.2.0
accelerate>=0.30
llama-cpp-python>=0.2.85
